Refactor import paths for useIsMobile hook across multiple components to improve code organization and maintainability.
