+++ app/(afterLogin)/(compensation)/deduction/allDeduction/_components/allDeductionTable.tsx
@@ -6,7 +6,7 @@ import { EmployeeDetails } from '../../../_components/employeeDetails';
 import { useAllAllowanceStore } from '@/store/uistate/features/compensation/allowance';
 import CustomPagination from '@/components/customPagination';
 import { CustomMobilePagination } from '@/components/customPagination/mobilePagination';
-import { useIsMobile } from '@/components/common/hooks/useIsMobile';
+import { useIsMobile } from '@/hooks/useIsMobile';
 
 const AllDeductionTable = () => {
   const { data: allCompensationsData, isLoading } = useFetchAllowances();
